==> MorrisCountyImages/01_getMorrisAddress.ipynb <==
### This file gets a list of addresses in Morris county by neighborhood

import pandas as pd
import numpy as np
from shapely.geometry import Point
import geopandas as gpd
from tqdm import tqdm

req_cols = ['FULLADDR','POST_CODE','POST_COMM','COUNTY','PLACE_TYPE','PLACEMENT','LONG_', 'LAT']
data = gpd.read_file('../Data/Addr_NG911.gdb', include_fields=req_cols) # 5m 51s for full df

### 1.) Subset initial address dataset to Morris County only

df = data.drop("geometry", axis=1).copy()
df["PLACE_TYPE"] = df["PLACE_TYPE"].mask(~df["PLACE_TYPE"].isin(['Residence', 'Unknown', 'Industrial', 'Outdoors']), "Commercial") # Sets all other possible Place Types to commercial as it fits that category
df.dropna(inplace=True)
df['COUNTY'].isnull().value_counts()
df["COUNTY"] = df["COUNTY"].astype(np.int64) # need to convert this column into int64
df = df[df["COUNTY"]==882231] # Morris County
df["ADDRESS"] = df["FULLADDR"] + ", " + df["POST_COMM"] + ", NJ " + df["POST_CODE"]
morris_add_list = df[["POST_COMM","PLACE_TYPE","ADDRESS","LONG_","LAT"]]

### 2.) Get town list from municipalities geofiles and save

filename = "../Data/morrisMunicipalities/Municipalities.shp"
gdf = gpd.read_file(filename).reset_index()
town_list = []

for index in range(39):
    row = gdf[gdf["index"] == index]
    town = row.Label.values[0]
    town_list.append(town)

towns = pd.DataFrame({'towns': town_list})
towns.to_csv("./csv/morris_towns.csv", index=False)

### 3.) Using a dict to try to replace as many town names as possible before the problem

# History for townships needed to be researched as some neighbourhoods were represented by different names, for instance, 
# long valley is a census-designated place (CDP) and most addresses in Washington Twp are in Long Valley.
 # 'Dover Town' --> Dover
 #'Chatham Township',
 #'Rockaway Borough',
 #'Mendham Borough',
 #'Mount Olive Township', --> budd lake
 #'Morristown Town',
 #'Washington Township', --> long valley
 #'Mount Arlington Borough',
 #'Chester Township',
 #'Montville Township',
 #'Parsippany-Troy Hills Township',
 #'Butler Borough',
 #'Long Hill Township', --> 'Stirling','Millington','Gillette'
 #'Victory Gardens Borough', --> Dover (small section, usually put as dover)
 #'Jefferson Township', --> 'Oak Ridge', 'Lake Hopatcong'
 #'Chester Borough',
 #'Mountain Lakes Borough',
 #'Denville Township',
 #'Mendham Township',
 #'Madison Borough',
 #'Pequannock Township',
 #'Hanover Township', --> 'Cedar Knolls','Whippany'
 #'Lincoln Park Borough',
 #'Kinnelon Borough',
 #'Morris Plains Borough',
 #'Netcong Borough',
 #'Harding Township', --> 'Green Village','New Vernon'
 #'Chatham Borough',
 #'Roxbury Township', --> 'Ledgewood', 'Succasunna', 'Kenvil'
 #'Morris Township',
 #'Wharton Borough',
 #'Boonton Town',
 #'Florham Park Borough',
 #'Randolph Township',
 #'Boonton Township',
 #'Rockaway Township',
 #'Mine Hill Township',
 #'East Hanover Township',
 #'Riverdale'
 
problems = ['Chatham Borough','Chatham Township',
        'Boonton Town', 'Boonton Township',
        'Mendham Borough', 'Mendham Township',
        'Rockaway Borough', 'Rockaway Township',
        'Chester Borough', 'Chester Township',
        'Morristown Town', 'Morris Township',
        'Dover Town', 'Victory Gardens Borough']
non_problems =  list(set(towns) - set(problems))

some_dict = {'Randolph':'Randolph Township','Madison':'Madison Borough','Parsippany':'Parsippany-Troy Hills Township','Kinnelon':'Kinnelon Borough','Morris Plains':'Morris Plains Borough','Butler':'Butler Borough','Denville':'Denville Township','East Hanover':'East Hanover Township','Montville':'Montville Township','Wharton':'Wharton Borough','Lincoln Park':'Lincoln Park Borough','Florham Park':'Florham Park Borough','Mine Hill':'Mine Hill Township','Netcong':'Netcong Borough','Mountain Lakes':'Mountain Lakes Borough','Mount Arlington':'Mount Arlington Borough','Pequannock':'Pequannock Township','Budd Lake':'Mount Olive Township','Long Valley':'Washington Township','Riverdale':'Riverdale Borough','Stirling':'Long Hill Township','Millington':'Long Hill Township','Gillette':'Long Hill Township','Oak Ridge':'Jefferson Township','Lake Hopatcong':'Jefferson Township','Cedar Knolls':'Hanover Township','Whippany':'Hanover Township','Green Village':'Harding Township','New Vernon':'Harding Township','Ledgewood':'Roxbury Township','Succasunna':'Roxbury Township','Kenvil':'Roxbury Township','Chatham':'Chatham','Boonton':'Boonton','Mendham':'Mendham','Rockaway':'Rockaway','Chester':'Chester','Morristown':'Morristown','Dover':'Dover'}

p = ['Chatham','Boonton','Mendham','Rockaway','Chester','Morristown', 'Dover']

morris_add = morris_add_list.copy()
morris_add.replace({'POST_COMM': some_dict}, inplace=True)

df_good = morris_add[morris_add.POST_COMM.isin(non_problems)].reset_index(drop=True)
df_bad = morris_add[morris_add.POST_COMM.isin(p)].reset_index(drop=True)

### 4.) Now we need to sort by coordinates to identify towns

df_bad

gdf = gdf.to_crs(4326) # Change coordinate systems
bounds = [gdf.iloc[[x]].unary_union.bounds for x in range(39)]
gdf = gdf[gdf["Label"].isin(problems)].reset_index(drop=True)
gdf

# 12 mins
new_town_list = []
for row in tqdm(range(len(df_bad))):
    row_add = df_bad.iloc[[row]]
    x = row_add.LONG_
    y = row_add.LAT
    p = Point(x, y)
    
    town_var = row_add.POST_COMM.values[0]
    for index in range(len(gdf)):
        gdf_row = gdf.iloc[[index]]
        #print(gdf_row)
        town = gdf_row.Label.values[0] 
        
        aoi_geom = gdf_row.unary_union
        
        if aoi_geom.contains(p):
            town_var = town
            break
    new_town_list.append(town_var)

new_town_list

new_df_bad = df_bad.copy()
new_df_bad.POST_COMM = new_town_list
new_df_bad = new_df_bad[new_df_bad["POST_COMM"].isin(towns)]
combined_df = pd.concat([df_good, new_df_bad], ignore_index=True)
combined_df.POST_COMM.nunique()

combined_df

combined_df.to_csv('./csv/morris_address.csv', index=False)
